github_app/requirements.txt
flask>=3.0
flask-cors>=4.0
requests>=2.31
PyJWT>=2.8
cryptography>=41.0
python-dotenv>=1.0
